Add support for angle bracket arguments and cite commands in particular.
Apacite uses the syntax \cite<text before citation>[text after citation]{bibEntries}, but the angle brackets cause latexdiff to mess up and the resulting diff doesn't compile.  Angle brackets have been added to the token regex as arguments for functions, as well as the regex that detects cite commands for protection with an mbox.